requirements.txt
python-telegram-bot==20.7
openpyxl
pandas==2.2.2
requests
pillow


